merged-project-song/producer/requirements.txt
requests==2.31.0
kafka-python==2.0.2
python-dotenv==1.0.0
pandas==2.0.3
